+++ src/i18n.js
@@ -38,6 +38,8 @@
     messages.ru.available_translations = 'Доступные переводы';
     messages.ru.voice_teams = 'Озвучка';
     messages.ru.subtitle_teams = 'Субтитры';
+    messages.ru.schedule_has_voice = 'На эту серию уже есть озвучка';
+    messages.ru.schedule_has_subtitles = 'На эту серию уже есть субтитры';
     messages.ru.translations_unknown = 'Информация о переводах пока недоступна';
     messages.ru.top_all = 'Общий топ';
     messages.ru.top_tv = 'Сериалы';
@@ -53,6 +55,8 @@
     messages.en.available_translations = 'Available translations';
     messages.en.voice_teams = 'Voice teams';
     messages.en.subtitle_teams = 'Subtitles';
+    messages.en.schedule_has_voice = 'A dub is already available for this episode';
+    messages.en.schedule_has_subtitles = 'Subtitles are already available for this episode';
     messages.en.translations_unknown = 'Translation information is not available yet';
     messages.en.version_label = 'Version';
     messages.en.top_all = 'Overall';
@@ -594,6 +598,8 @@
     messages.uk.available_translations = 'Доступні переклади';
     messages.uk.voice_teams = 'Озвучення';
     messages.uk.subtitle_teams = 'Субтитри';
+    messages.uk.schedule_has_voice = 'На цю серію вже є озвучення';
+    messages.uk.schedule_has_subtitles = 'На цю серію вже є субтитри';
     messages.uk.translations_unknown = 'Інформація про переклади поки недоступна';
     messages.uk.back_to_lampa = 'Повернутися до Lampa';
     messages.uk.top_all = 'Загальний топ';